+++ src/compas_timber/model.py
@@ -129,6 +129,18 @@ class TimberModel(Model):
         return candidates
 
     @property
+    def unpromoted_joint_candidates(self):
+        """returns JointCandidates that lie on edges that don't also have a Joint instance on them."""
+        # type: () -> set[JointCandidate]
+        unpromoted = set()
+        for edge in self._graph.edges():
+            edge_candidate = self._graph.edge_attribute(edge, "candidates")
+            edge_joint = self._graph.edge_attribute(edge, "joints")
+            if edge_candidate is not None and edge_joint is None:
+                unpromoted.add(edge_candidate)
+        return unpromoted
+
+    @property
     def topologies(self):
         return self._topologies